+++ octofludb/genbank.py
@@ -1,25 +1,54 @@
-from octofludb.nomenclature import uidgen, P, O, make_property, make_uri
+from octofludb.nomenclature import (
+    uidgen,
+    P,
+    O,
+    make_property,
+    make_uri,
+    make_date,
+    make_country_uri,
+    make_usa_state_uri,
+)
 from octofludb.util import rmNone, make_maybe_add
 from octofludb.hash import chksum
 from rdflib import Literal
+from rdflib.namespace import XSD
+import re
+import octofludb.domain.identifier as identifier
+import octofludb.domain.animal as animal
 import sys
+from octofludb.util import log
+from octofludb.colors import bad
+import octofludb.domain.geography as geo
 
 
-def add_gb_meta_triples(g, gb_meta):
+def add_gb_meta_triples(g, gb_meta, only_influenza_a=True):
+
+    if only_influenza_a:
+        # ignore this entry if the organism is not specified
+        if not "GBSeq_organism" in gb_meta:
+            return None
+        # ignore this entry if the organism is not an Influenza virus
+        if not bool(re.match("Influenza [ABCD] virus", gb_meta["GBSeq_organism"])):
+            return None
+
     accession = str(gb_meta["GBSeq_primary-accession"])
 
     gid = make_uri(accession)
     g.add((gid, P.gb, Literal(accession)))
 
     maybe_add = make_maybe_add(g, gb_meta, gid)
-    maybe_add(P.gb_length, "GBSeq_locus")
-    maybe_add(P.gb_length, "GBSeq_length")
+
+    def make_integer(x):
+        return rdflib.Literal(x, datatype=XSD.integer)
+
+    maybe_add(P.gb_locus, "GBSeq_locus")
+    maybe_add(P.gb_length, "GBSeq_length", formatter=make_integer)
     maybe_add(P.gb_strandedness, "GBSeq_strandedness")
     maybe_add(P.gb_moltype, "GBSeq_moltype")
     maybe_add(P.gb_topology, "GBSeq_topology")
     maybe_add(P.gb_division, "GBSeq_division")
-    maybe_add(P.gb_update_date, "GBSeq_update-date")
-    maybe_add(P.gb_create_date, "GBSeq_create-date")
+    maybe_add(P.gb_update_date, "GBSeq_update-date", formatter=make_date)
+    maybe_add(P.gb_create_date, "GBSeq_create-date", formatter=make_date)
     maybe_add(P.gb_definition, "GBSeq_definition")
     maybe_add(P.gb_primary_accession, "GBSeq_primary_accession")
     maybe_add(P.gb_accession_version, "GBSeq_accession-version")
@@ -27,9 +56,16 @@ def add_gb_meta_triples(g, gb_meta):
     maybe_add(P.gb_organism, "GBSeq_organism")
     maybe_add(P.gb_taxonomy, "GBSeq_taxonomy")
 
-    seq = gb_meta["GBSeq_sequence"].upper()
-    g.add((gid, P.dnaseq, Literal(seq)))
-    g.add((gid, P.chksum, Literal(chksum(seq))))
+    # usually an entry has sequence, but there are weird exceptions
+    if "GBSeq_sequence" in gb_meta:
+        seq = gb_meta["GBSeq_sequence"].upper()
+        g.add((gid, P.dnaseq, Literal(seq)))
+        g.add((gid, P.chksum, Literal(chksum(seq))))
+
+    strain = None
+    host = None
+    date = None
+    country = None
 
     igen = uidgen(base=accession + "_feat_")
     for feat in gb_meta["GBSeq_feature-table"]:
@@ -43,15 +79,53 @@ def add_gb_meta_triples(g, gb_meta):
 
         if "GBFeature_quals" in feat:
             for qual in feat["GBFeature_quals"]:
-                if qual["GBQualifier_name"] == "translation":
-                    aaseq = qual["GBQualifier_value"]
-                    g.add((fid, P.proseq, Literal(aaseq)))
-                    g.add((fid, P.chksum, Literal(chksum(aaseq))))
-                elif "GBQualifier_name" in qual and "GBQualifier_value" in qual:
-                    g.add(
-                        (
-                            fid,
-                            make_property(qual["GBQualifier_name"]),
-                            Literal(qual["GBQualifier_value"]),
-                        )
-                    )
+                if not ("GBQualifier_name" in qual and "GBQualifier_value" in qual):
+                    continue
+                key = qual["GBQualifier_name"]
+                val = qual["GBQualifier_value"]
+
+                if key == "translation":
+                    g.add((fid, P.proseq, Literal(val)))
+                    g.add((fid, P.chksum, Literal(chksum(val))))
+                elif key == "strain":
+                    try:
+                        strain = identifier.p_strain.parse(val)
+                    except:
+                        log(bad("Bad strain name: ") + val)
+                        strain = val
+                elif key == "collection_date":
+                    date = make_date(val)
+                elif key == "host":
+                    host = val
+                elif key == "country":
+                    country = re.sub(":.*", "", val)
+                else:
+                    g.add((fid, make_property(key), Literal(val)))
+
+    # link strain information
+    if strain:
+        sid = make_uri(strain)
+        g.add((sid, P.has_segment, gid))
+        g.add((sid, P.strain_name, Literal(strain)))
+        if host:
+            g.add((sid, P.host, Literal(animal.clean_host(host))))
+        if date:
+            g.add((sid, P.date, date))
+        if country:
+            code = geo.country_to_code(country)
+            g.add((sid, P.country, make_country_uri(code)))
+            if code == "USA":
+                fields = strain.split("/")
+                for field in fields[1:]:
+                    # If this looks like a US state, add it
+                    code = geo.state_to_code(field)
+                    if code:
+                        g.add((sid, P.state, make_usa_state_uri(code)))
+                    # If this looks like an A0 number, add it
+                    try:
+                        A0 = identifier.p_A0.parse_strict(field)
+                        g.add((sid, P.barcode, Literal(A0)))
+                    except:
+                        pass
+    else:
+        log(bad("Missing strain: ") + gb_meta["GBSeq_locus"])